Move PerformanceStatisticRow into tooltip class
